interfaces: udp: Change comparison target
In this commit, the comparison target
of received_len was changed to header-size.
In the previous case,
csp_if_udp_rx_work does not work properly
because CSP v1 and v2 have different header-size.

A literal number 4 is used for both protocol v1 and v2.
For v1, the header size is 4, but for v2, it’s 6.

Signed-off-by: Riho Natsushima <nasshi7224@gmail.com>